+++ java/Ice/greeter/server/src/main/java/com/example/ice/greeter/server/Server.java
@@ -1,6 +1,6 @@
 // Copyright (c) ZeroC, Inc.
 
-package com.zeroc.demos.ice.greeter.serveramd;
+package com.example.ice.greeter.server;
 
 import com.zeroc.Ice.Communicator;
 import com.zeroc.Ice.ObjectAdapter;
