+++ java/Ice/greeter/serveramd/src/main/java/com/example/ice/greeter/serveramd/Server.java
@@ -1,6 +1,6 @@
 // Copyright (c) ZeroC, Inc.
 
-package com.zeroc.demos.ice.greeter.server;
+package com.example.ice.greeter.serveramd;
 
 import com.zeroc.Ice.Communicator;
 import com.zeroc.Ice.ObjectAdapter;
